Move tests to proper file
