+++ tests/test_speed_optimizations.py
@@ -1,3 +1,4 @@
+import numpy as np
 import pytest
 import torch
 
@@ -133,9 +134,11 @@ def test_inactive_magnet_is_replaced_by_drift():
     """
     segment = cheetah.Segment(
         elements=[
-            cheetah.Drift(length=torch.tensor([0.6])),
-            cheetah.Quadrupole(length=torch.tensor([0.2]), k1=torch.tensor([0.0])),
-            cheetah.Drift(length=torch.tensor([0.4])),
+            cheetah.Drift(length=torch.tensor([0.6, 0.5])),
+            cheetah.Quadrupole(
+                length=torch.tensor([0.2, 0.3]), k1=torch.tensor([0.0, 0.0])
+            ),
+            cheetah.Drift(length=torch.tensor([0.4, 0.1])),
         ]
     )
 
@@ -144,6 +147,7 @@ def test_inactive_magnet_is_replaced_by_drift():
     assert all(
         isinstance(element, cheetah.Drift) for element in optimized_segment.elements
     )
+    assert np.allclose(segment.length, optimized_segment.length)
 
 
 def test_active_elements_not_replaced_by_drift():
@@ -152,9 +156,11 @@ def test_active_elements_not_replaced_by_drift():
     """
     segment = cheetah.Segment(
         elements=[
-            cheetah.Drift(length=torch.tensor([0.6])),
-            cheetah.Quadrupole(length=torch.tensor([0.2]), k1=torch.tensor([4.2])),
-            cheetah.Drift(length=torch.tensor([0.4])),
+            cheetah.Drift(length=torch.tensor([0.6, 0.5])),
+            cheetah.Quadrupole(
+                length=torch.tensor([0.2, 0.3]), k1=torch.tensor([4.2, 0.0])
+            ),
+            cheetah.Drift(length=torch.tensor([0.4, 0.1])),
         ]
     )
 
@@ -182,6 +188,25 @@ def test_inactive_magnet_drift_replacement_dtype(dtype: torch.dtype):
     optimized_segment = segment.inactive_elements_as_drifts()
 
     assert all(element.length.dtype == dtype for element in optimized_segment.elements)
+
+
+def test_inactive_magnet_drift_except_for():
+    """
+    Test that an inactive magnet is not replaced by a drift if except is used
+    """
+    segment = cheetah.Segment(
+        elements=[
+            cheetah.Drift(length=torch.tensor([0.6])),
+            cheetah.Quadrupole(
+                length=torch.tensor([0.2]), k1=torch.tensor([4.2]), name="quad"
+            ),
+            cheetah.Drift(length=torch.tensor([0.4])),
+        ]
+    )
+
+    optimized_segment = segment.inactive_elements_as_drifts(except_for=["quad"])
+
+    assert isinstance(optimized_segment.elements[1], cheetah.Quadrupole)
 
 
 def test_skippable_elements_reset():
@@ -214,3 +239,39 @@ def test_skippable_elements_reset():
     merged_tm = merged_segment.elements[2].transfer_map(energy=incoming_beam.energy)
 
     assert torch.allclose(original_tm, merged_tm)
+
+
+def test_without_zerolength_elements():
+    """Test that zerolength elements are properly recognized and removed."""
+    segment = cheetah.Segment(
+        elements=[
+            cheetah.Drift(length=torch.tensor([1.0, 2.0])),
+            cheetah.Dipole(
+                length=torch.tensor([0.0, 0.0]),
+                angle=torch.tensor([0.0, 0.0]),
+            ),
+            cheetah.Dipole(
+                length=torch.tensor([0.0, 0.0]),
+                angle=torch.tensor([0.0, 0.0]),
+                name="dipole",
+            ),
+            cheetah.Dipole(
+                length=torch.tensor([0.0, 0.1]),
+                angle=torch.tensor([0.0, 0.0]),
+            ),
+            cheetah.Drift(length=torch.tensor([0.0, 0.0])),
+            cheetah.Dipole(
+                length=torch.tensor([0.0, 0.0]),
+                angle=torch.tensor([0.5, 0.0]),
+            ),
+        ]
+    )
+
+    pruned = segment.without_inactive_zero_length_elements()
+    pruned_except = segment.without_inactive_zero_length_elements(except_for=["dipole"])
+
+    assert len(segment.elements) == 6
+    assert len(pruned.elements) == 3
+    assert len(pruned_except.elements) == 4
+    assert np.allclose(segment.length, pruned.length)
+    assert np.allclose(segment.length, pruned_except.length)
